+++ core/src/main/java/bisq/core/trade/messages/mediation/MediatedPayoutTxPublishedMessage.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.trade.messages.dispute;
+package bisq.core.trade.messages.mediation;
 
 import bisq.core.trade.messages.TradeMessage;
 
@@ -23,27 +23,26 @@ import bisq.network.p2p.MailboxMessage;
 import bisq.network.p2p.NodeAddress;
 
 import bisq.common.app.Version;
+import bisq.common.proto.network.NetworkEnvelope;
 import bisq.common.util.Utilities;
 
 import com.google.protobuf.ByteString;
 
 import lombok.EqualsAndHashCode;
 import lombok.Value;
-import lombok.extern.slf4j.Slf4j;
 
-@Slf4j
-@Value
 @EqualsAndHashCode(callSuper = true)
-public class MediatedPayoutTxSignatureMessage extends TradeMessage implements MailboxMessage {
-    private final byte[] txSignature;
+@Value
+public final class MediatedPayoutTxPublishedMessage extends TradeMessage implements MailboxMessage {
+    private final byte[] payoutTx;
     private final NodeAddress senderNodeAddress;
 
-    public MediatedPayoutTxSignatureMessage(byte[] txSignature,
-                                            String tradeId,
+    public MediatedPayoutTxPublishedMessage(String tradeId,
+                                            byte[] payoutTx,
                                             NodeAddress senderNodeAddress,
                                             String uid) {
-        this(txSignature,
-                tradeId,
+        this(tradeId,
+                payoutTx,
                 senderNodeAddress,
                 uid,
                 Version.getP2PMessageVersion());
@@ -54,48 +53,41 @@ public class MediatedPayoutTxSignatureMessage extends TradeMessage implements Ma
     // PROTO BUFFER
     ///////////////////////////////////////////////////////////////////////////////////////////
 
-    private MediatedPayoutTxSignatureMessage(byte[] txSignature,
-                                             String tradeId,
+    private MediatedPayoutTxPublishedMessage(String tradeId,
+                                             byte[] payoutTx,
                                              NodeAddress senderNodeAddress,
                                              String uid,
                                              int messageVersion) {
         super(messageVersion, tradeId, uid);
-        this.txSignature = txSignature;
+        this.payoutTx = payoutTx;
         this.senderNodeAddress = senderNodeAddress;
     }
 
     @Override
     public protobuf.NetworkEnvelope toProtoNetworkEnvelope() {
         return getNetworkEnvelopeBuilder()
-                .setMediatedPayoutTxSignatureMessage(protobuf.MediatedPayoutTxSignatureMessage.newBuilder()
-                        .setTxSignature(ByteString.copyFrom(txSignature))
+                .setMediatedPayoutTxPublishedMessage(protobuf.MediatedPayoutTxPublishedMessage.newBuilder()
                         .setTradeId(tradeId)
+                        .setPayoutTx(ByteString.copyFrom(payoutTx))
                         .setSenderNodeAddress(senderNodeAddress.toProtoMessage())
                         .setUid(uid))
                 .build();
     }
 
-    public static MediatedPayoutTxSignatureMessage fromProto(protobuf.MediatedPayoutTxSignatureMessage proto,
-                                                             int messageVersion) {
-        return new MediatedPayoutTxSignatureMessage(proto.getTxSignature().toByteArray(),
-                proto.getTradeId(),
+    public static NetworkEnvelope fromProto(protobuf.MediatedPayoutTxPublishedMessage proto, int messageVersion) {
+        return new MediatedPayoutTxPublishedMessage(proto.getTradeId(),
+                proto.getPayoutTx().toByteArray(),
                 NodeAddress.fromProto(proto.getSenderNodeAddress()),
                 proto.getUid(),
                 messageVersion);
     }
 
     @Override
-    public String getTradeId() {
-        return tradeId;
-    }
-
-
-    @Override
     public String toString() {
-        return "MediatedPayoutSignatureMessage{" +
-                "\n     txSignature=" + Utilities.bytesAsHexString(txSignature) +
-                ",\n     tradeId='" + tradeId + '\'' +
+        return "MediatedPayoutTxPublishedMessage{" +
+                "\n     payoutTx=" + Utilities.bytesAsHexString(payoutTx) +
                 ",\n     senderNodeAddress=" + senderNodeAddress +
+                ",\n     uid='" + uid + '\'' +
                 "\n} " + super.toString();
     }
 }
