+++ core/src/main/java/bisq/core/trade/protocol/TradeProtocol.java
@@ -25,8 +25,8 @@ import bisq.core.trade.messages.InputsForDepositTxRequest;
 import bisq.core.trade.messages.PeerPublishedDelayedPayoutTxMessage;
 import bisq.core.trade.messages.TradeMessage;
 import bisq.core.trade.messages.TraderSignedWitnessMessage;
-import bisq.core.trade.messages.dispute.MediatedPayoutTxPublishedMessage;
-import bisq.core.trade.messages.dispute.MediatedPayoutTxSignatureMessage;
+import bisq.core.trade.messages.mediation.MediatedPayoutTxPublishedMessage;
+import bisq.core.trade.messages.mediation.MediatedPayoutTxSignatureMessage;
 import bisq.core.trade.protocol.tasks.ApplyFilter;
 import bisq.core.trade.protocol.tasks.ProcessPeerPublishedDelayedPayoutTxMessage;
 import bisq.core.trade.protocol.tasks.mediation.BroadcastMediatedPayoutTx;
