+++ core/src/main/java/bisq/core/trade/protocol/tasks/mediation/ProcessMediatedPayoutSignatureMessage.java
@@ -19,7 +19,7 @@ package bisq.core.trade.protocol.tasks.mediation;
 
 import bisq.core.support.dispute.mediation.MediationResultState;
 import bisq.core.trade.Trade;
-import bisq.core.trade.messages.dispute.MediatedPayoutTxSignatureMessage;
+import bisq.core.trade.messages.mediation.MediatedPayoutTxSignatureMessage;
 import bisq.core.trade.protocol.tasks.TradeTask;
 import bisq.core.util.Validator;
