+++ core/src/main/java/bisq/core/trade/protocol/tasks/mediation/SendMediatedPayoutSignatureMessage.java
@@ -20,7 +20,7 @@ package bisq.core.trade.protocol.tasks.mediation;
 import bisq.core.support.dispute.mediation.MediationResultState;
 import bisq.core.trade.Contract;
 import bisq.core.trade.Trade;
-import bisq.core.trade.messages.dispute.MediatedPayoutTxSignatureMessage;
+import bisq.core.trade.messages.mediation.MediatedPayoutTxSignatureMessage;
 import bisq.core.trade.protocol.tasks.TradeTask;
 
 import bisq.network.p2p.NodeAddress;
